Handle possible "race condition" with on_joinplayer
Sometimes when on_joinplayer is called the player does not entirely
exist yet, so the result of `player:get_player_name()` is an empty
string. Due to this, get_player_information returns nil, and when
chat3 attempts to access and cache the protocol version from what
it expects to be a table, the game crashes.

This commit solves the issue by deferring caching until the first
time the protocol version is requested if the player does not exist
at the time of on_joinplayer being run.

If a player were to send a chat message before they existed it
would fail to send due to the way this solution works. However,
Minetest would have to fail horribly for this to happen in the
first place.

Should resolve #3.